+++ pakyow-core/spec/integration/runnable/container/success_spec.rb
@@ -178,4 +178,12 @@ RSpec.describe "determining container success" do
 
     include_examples :examples
   end
+
+  context "hybrid container" do
+    let(:run_options) {
+      { strategy: :hybrid }
+    }
+
+    include_examples :examples
+  end
 end